Fix: Health Operation
This is the final fix before I claim it as version 0.1
(づ′▽`)づ

Allow to replace the origin char after the ball or the star blocking it, not a space.

Replace different position struct to position, in system_utils. I remember there is a ball_pos struct in physic

Midground and background x and y not represent same as other, now united, down-left in print is 0,0.

Improve physic logic, especially the fixed flag , I add a platform on the air for testing, it runs well~