Feat: Dialog Box
A big test for the invoke function,  a dialog box with a 5-second hanging. The ball remains active when the dialog box is up, and the box disappears when the ball stops or bouncing. Good!
Also, I add recover_background_batch to clean this large dialog box.

BTW, there is a anime girl, she is by the dialog box, in ascii art.